03/10/2022 - Temporary final commit
This is the final commit for this project for now.
I would like to come back to it since I liked the
concept of the website a lot as I was creating it.
I want to add animations, interactions, and use
api once I get more in deapth into JavaScript.